Refine favorites code
